refactor: seperate database access code
